Add tests for new fulltext API
