+++ packages/graphql/src/schema/resolvers/query/aggregate.ts
@@ -23,7 +23,6 @@ import type { Node } from "../../../classes";
 import type { Context } from "../../../types";
 import { translateAggregate } from "../../../translate";
 import getNeo4jResolveTree from "../../../utils/get-neo4j-resolve-tree";
-import { createFulltextDeprecationMessage } from "../../../schema/augment/fulltext";
 
 export function aggregateResolver({ node }: { node: Node }) {
     async function resolve(_root: any, _args: any, _context: unknown, info: GraphQLResolveInfo) {
@@ -57,14 +56,6 @@ export function aggregateResolver({ node }: { node: Node }) {
                 ? {
                       fulltext: {
                           type: `${node.name}Fulltext`,
-                          directives: [
-                              {
-                                  name: "deprecated",
-                                  args: {
-                                      reason: createFulltextDeprecationMessage(node),
-                                  },
-                              },
-                          ],
                       },
                   }
                 : {}),
