+++ packages/graphql/tests/schema/fulltext.test.ts
@@ -198,7 +198,7 @@ describe("@fulltext schema", () => {
 
             type Query {
               movies(fulltext: MovieFulltext @deprecated(reason: \\"This argument has been deprecated and will be removed in version 4.0.0 of the library.\\\\nPlease use the fulltext argument in moviesConnection with index you wish to query instead.\\\\nMore information about the changes to @fulltext can be found here:\\\\nhttps://neo4j.com/docs/graphql-manual/current/guides/v4-migration/#_fulltext_changes.\\\\n\\"), options: MovieOptions, where: MovieWhere): [Movie!]!
-              moviesAggregate(fulltext: MovieFulltext @deprecated(reason: \\"This argument has been deprecated and will be removed in version 4.0.0 of the library.\\\\nPlease use the fulltext argument in moviesConnection with index you wish to query instead.\\\\nMore information about the changes to @fulltext can be found here:\\\\nhttps://neo4j.com/docs/graphql-manual/current/guides/v4-migration/#_fulltext_changes.\\\\n\\"), where: MovieWhere): MovieAggregateSelection!
+              moviesAggregate(fulltext: MovieFulltext, where: MovieWhere): MovieAggregateSelection!
               moviesConnection(after: String, first: Int, fulltext: MovieFulltext, sort: [MovieSort], where: MovieWhere): MoviesConnection!
               moviesFulltextMovieDescription(limit: Int, offset: Int, phrase: String!, sort: [MovieFulltextSort!], where: MovieFulltextWhere): [MovieFulltextResult!]! @deprecated(reason: \\"This operation has been deprecated and will be removed in version 4.0.0 of the library.\\\\nPlease use the fulltext argument in moviesConnection with index you wish to query instead.\\\\nMore information about the changes to @fulltext can be found here:\\\\nhttps://neo4j.com/docs/graphql-manual/current/guides/v4-migration/#_fulltext_changes.\\\\n\\")
               moviesFulltextMovieTitle(limit: Int, offset: Int, phrase: String!, sort: [MovieFulltextSort!], where: MovieFulltextWhere): [MovieFulltextResult!]! @deprecated(reason: \\"This operation has been deprecated and will be removed in version 4.0.0 of the library.\\\\nPlease use the fulltext argument in moviesConnection with index you wish to query instead.\\\\nMore information about the changes to @fulltext can be found here:\\\\nhttps://neo4j.com/docs/graphql-manual/current/guides/v4-migration/#_fulltext_changes.\\\\n\\")
